pyudmi/pyproject.toml
[project]
name = "pyudmi"
version = "0.1.0"
description = "A Python library for UDMI"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
udmi-schema = {path = "../gencode/python/dist/udmi_schema-0.1.0-py3-none-any.whl"}
paho-mqtt = ">=2.1.0,<3.0.0"
PyJWT = "^2.8.0"
cryptography = "^41.0.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-mock = "^3.15.1"
freezegun = "^1.5.5"

[tool.poetry]
packages = [
    { include = "udmi", from = "src" },
]
